requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-axes==5.35.0
django-crontab==0.7.1
django-ipware==4.0.2
djangorestframework==3.13.1
exponent-server-sdk==2.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
openapi-codec==1.3.2
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
